refactor taskmanager into setup and register functions
